feat(site): bring back dark blue
