+++ site/src/components/Dashboard/LicenseBanner/LicenseBannerView.tsx
@@ -1,15 +1,14 @@
+import Link from "@mui/material/Link";
 import {
+  css,
   type CSSObject,
   type Interpolation,
   type Theme,
   useTheme,
 } from "@emotion/react";
-import Link from "@mui/material/Link";
-import { css } from "@emotion/react";
-import { useState } from "react";
+import { type FC, useState } from "react";
 import { Expander } from "components/Expander/Expander";
 import { Pill } from "components/Pill/Pill";
-import { colors } from "theme/colors";
 
 export const Language = {
   licenseIssue: "License Issue",
@@ -32,7 +31,7 @@ export interface LicenseBannerViewProps {
   warnings: string[];
 }
 
-export const LicenseBannerView: React.FC<LicenseBannerViewProps> = ({
+export const LicenseBannerView: FC<LicenseBannerViewProps> = ({
   errors,
   warnings,
 }) => {
@@ -48,7 +47,9 @@ export const LicenseBannerView: React.FC<LicenseBannerViewProps> = ({
     display: flex;
     align-items: center;
     padding: 12px;
-    background-color: ${type === "error" ? colors.red[10] : colors.orange[10]};
+    background-color: ${type === "error"
+      ? theme.colors.red[10]
+      : theme.colors.orange[10]};
   `;
 
   if (messages.length === 1) {
