+++ site/src/components/Dashboard/Navbar/UserDropdown/UserDropdown.tsx
@@ -1,12 +1,11 @@
 import Badge from "@mui/material/Badge";
-import { type FC, type PropsWithChildren } from "react";
-import { colors } from "theme/colors";
+import { type FC, type ReactNode } from "react";
 import type * as TypesGen from "api/typesGenerated";
 import { BUTTON_SM_HEIGHT, navHeight } from "theme/constants";
 import { DropdownArrow } from "components/DropdownArrow/DropdownArrow";
 import { UserAvatar } from "components/UserAvatar/UserAvatar";
 import { UserDropdownContent } from "./UserDropdownContent";
-import { css } from "@emotion/react";
+import { css, type Interpolation, type Theme, useTheme } from "@emotion/react";
 import {
   Popover,
   PopoverContent,
@@ -19,67 +18,49 @@ export interface UserDropdownProps {
   supportLinks?: TypesGen.LinkConfig[];
   onSignOut: () => void;
   isDefaultOpen?: boolean;
+  children?: ReactNode;
 }
 
-export const UserDropdown: FC<PropsWithChildren<UserDropdownProps>> = ({
+export const UserDropdown: FC<UserDropdownProps> = ({
   buildInfo,
   user,
   supportLinks,
   onSignOut,
   isDefaultOpen,
-}: UserDropdownProps) => {
+}) => {
+  const theme = useTheme();
+
   return (
     <Popover isDefaultOpen={isDefaultOpen}>
       {(popover) => (
         <>
           <PopoverTrigger>
-            <button
-              css={css`
-                background: none;
-                border: 0;
-                cursor: pointer;
-                height: ${navHeight}px;
-                padding: 12px 0;
-
-                &:hover {
-                  background-color: transparent;
-                }
-              `}
-              data-testid="user-dropdown-trigger"
-            >
-              <div
-                css={{
-                  display: "flex",
-                  alignItems: "center",
-                  minWidth: 0,
-                  maxWidth: 300,
-                }}
-              >
+            <button css={styles.button} data-testid="user-dropdown-trigger">
+              <div css={styles.badgeContainer}>
                 <Badge overlap="circular">
                   <UserAvatar
-                    css={{
-                      width: BUTTON_SM_HEIGHT,
-                      height: BUTTON_SM_HEIGHT,
-                      fontSize: 16,
-                    }}
+                    css={styles.avatar}
                     username={user.username}
                     avatarURL={user.avatar_url}
                   />
                 </Badge>
-                <DropdownArrow color={colors.gray[6]} close={popover.isOpen} />
+                <DropdownArrow
+                  color={theme.colors.gray[6]}
+                  close={popover.isOpen}
+                />
               </div>
             </button>
           </PopoverTrigger>
 
           <PopoverContent
             horizontal="right"
-            css={(theme) => ({
+            css={{
               ".MuiPaper-root": {
                 minWidth: "auto",
                 width: 260,
                 boxShadow: theme.shadows[6],
               },
-            })}
+            }}
           >
             <UserDropdownContent
               user={user}
@@ -93,3 +74,30 @@ export const UserDropdown: FC<PropsWithChildren<UserDropdownProps>> = ({
     </Popover>
   );
 };
+
+const styles = {
+  button: css`
+    background: none;
+    border: 0;
+    cursor: pointer;
+    height: ${navHeight}px;
+    padding: 12px 0;
+
+    &:hover {
+      background-color: transparent;
+    }
+  `,
+
+  badgeContainer: {
+    display: "flex",
+    alignItems: "center",
+    minWidth: 0,
+    maxWidth: 300,
+  },
+
+  avatar: {
+    width: BUTTON_SM_HEIGHT,
+    height: BUTTON_SM_HEIGHT,
+    fontSize: 16,
+  },
+} satisfies Record<string, Interpolation<Theme>>;
