+++ site/src/theme/dark/index.ts
@@ -1,0 +1,9 @@
+import colors from "./colors";
+import experimental from "./experimental";
+import muiTheme from "./mui";
+
+export default {
+  ...muiTheme,
+  colors,
+  experimental,
+};
