+++ site/src/theme/darkBlue/colors.ts
@@ -1,0 +1,62 @@
+import tw from "../tailwind";
+
+export default {
+  white: "#fff",
+
+  gray: {
+    17: tw.slate[900],
+    16: tw.slate[900],
+    14: tw.slate[800],
+    13: tw.slate[700],
+    12: tw.slate[600],
+    11: tw.slate[500],
+    9: tw.slate[400],
+    6: tw.slate[300],
+    4: tw.slate[200],
+    2: tw.slate[100],
+    1: tw.slate[50],
+  },
+
+  red: {
+    15: tw.red[950],
+    12: tw.red[800],
+    10: tw.red[700],
+    9: tw.red[600],
+    8: tw.red[500],
+    6: tw.red[400],
+    2: tw.red[50],
+  },
+
+  orange: {
+    15: tw.amber[950],
+    14: tw.amber[900],
+    12: tw.amber[800],
+    11: tw.amber[700],
+    10: tw.amber[600],
+    9: tw.amber[500],
+    7: tw.amber[400],
+  },
+
+  yellow: {
+    5: tw.yellow[300],
+  },
+
+  green: {
+    15: tw.green[950],
+    13: tw.green[700],
+    12: tw.green[600],
+    11: tw.green[500],
+    9: tw.green[400],
+    8: tw.green[300],
+  },
+
+  blue: {
+    14: tw.blue[950],
+    9: tw.blue[600],
+    8: tw.blue[500],
+    7: tw.blue[400],
+    6: tw.blue[300],
+    3: tw.blue[200],
+    1: tw.blue[50],
+  },
+};
